It's working Yay, i put in the code what was the issue in the main function
