+++ coffee.txt
@@ -8,3 +8,5 @@ Randomly Named Coffee
 251.0
 adsff
 111.0
+Heather
+100.0
